pyproject.toml
[tool.poetry]
name = "jwt-manager"
version = "0.1.0"
description = ""
authors = ["vivien <bognetienoue@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"
redis = "^5.2.0"
pyjwt = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
